Add default font releaseGLObjects workaround for OSG <3.65
This resolves https://gitlab.com/OpenMW/openmw/issues/4704

I've not compiled this or anything and I'm going off my memory from
around ten or eleven months ago, so if this fails CI or causes more
problems, that's why. The most likely issue is that we'll need to make
the context current before releasing the program.