Install aws-cli
We have to get it from Amazon directly and run an install script as
aws-cli v2 (which is what we use on Windows) is only shipped as a snap,
and snaps don't work on Docker, which the GitLab runners use.

This is a combination of 4 commits.
This is the 1st commit message:

Install aws-cli

Get it from snap because apt only has 1.x versions.
Pass --classic so it can actually see files.

This is the commit message #2:

But snapd isn't installed by default even though it's Ubuntu's recommended way to install everything

This is the commit message #3:

Mysterious snap-fixing incantation

I guess the service isn't running my default until the next reboot?

This is the commit message #4:

Snap doesn't work in Docker, apt doesn't have awscli v2+, let's just install random binaries wherever they want to put themselves

    pick 5f0dfbf768 Install aws-cli
    squash 1ed2a711fc But snapd isn't installed by default even though it's Ubuntu's recommended way to install everything
    squash 230c917993 Mysterious snap-fixing incantation
    squash 75142ff3c2 Snap doesn't work in Docker, apt doesn't have awscli v2+, let's just install random binaries wherever they want to put themselves