Install the local openmw.cfg template on Windows
The Windows global config path is really really dumb and the whole concept of having a global path on an OS where it's so easy to have different incompatible versions installed side-by-side (instead of a single canonical OS-wide OpenMW install from the system package manager) is really dumb, too.

We've therefore only used local `openmw.cfg`s on Windows for a very long time.

However, it appears that we were installing the global `openmw.cfg` template instead of the local template.
That's wrong and we shouldn't do it, so I've stopped it.

I've left the separate `openmw.cfg` and `openmw.cfg.install` in the build directory as that makes doing development for portable installs with the same build system as doing the development for the installer easier.